Moved resizable propPanel to module
