+++ modules/propPanelResize/index.js
@@ -1,0 +1,6 @@
+import PropPanelResize from './PropPanelResize';
+
+export default {
+  __init__: ['propPanelResize'],
+  propPanelResize: ['type', PropPanelResize],
+};
